feat: move minio storage to doucment
